% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_constructors.R
\name{Blank}
\alias{Blank}
\title{Blank Curve constructor function}
\usage{
Blank()
}
\description{
This creates a Curve object for a 'Blank' pseudo-distribution.\cr
Curve objects contain all necessary information to describe a distribution, including functions and parameters describing it.\cr
This distribution is 0 by definition for all times. It is not therefore a true probability distribution.
}
\details{
The blank pseudo-distribution is used for impossible events, notably where censoring is not possible/allowed.\cr
f(x) = 0\cr
F(x) = 0
}
\examples{
Blank()
}
\author{
James Bell
}
