% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{createRFfunction,Curve-method}
\alias{createRFfunction,Curve-method}
\title{Method for creating a random draw function from a Curve object}
\usage{
\S4method{createRFfunction}{Curve}(theObject, n = "n")
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{n}{The parameter name for the number of random draws to make. Default=n}
}
\description{
This creates a random draw function from the Curve object
}
\examples{
createRFfunction(Weibull(100,1))
}
