% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{createRFfunction}
\alias{createRFfunction}
\title{Method for running the RF function for a Curve object}
\usage{
createRFfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{...}{Pass-through arguments}
}
\description{
This retrieves and runs the full RF function of the Curve object as a string
}
\examples{
createRFfunction(Weibull(100,1))
}
