% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{evaluateCDFfunction,Curve-method}
\alias{evaluateCDFfunction,Curve-method}
\title{Method for evaluating the CDF function for a Curve object at q}
\usage{
\S4method{evaluateCDFfunction}{Curve}(theObject, q)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{q}{The time to evaluate at}
}
\description{
This numerically evaluates the CDF function of the Curve object at the specified q
}
\examples{
evaluateCDFfunction(Weibull(100,1),10)
}
