% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{evaluateCDFfunction}
\alias{evaluateCDFfunction}
\title{Method for evaluating the CDF function for a Curve object at q}
\usage{
evaluateCDFfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{...}{Pass-through arguments}
}
\description{
This numerically evaluates the CDF function of the Curve object at the specified q
}
\examples{
evaluateCDFfunction(Weibull(100,1),10)
}
