% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{evaluateInvfunction,Curve-method}
\alias{evaluateInvfunction,Curve-method}
\title{Method for evaluating the inverse-CDF function for a Curve object at p}
\usage{
\S4method{evaluateInvfunction}{Curve}(theObject, p)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{p}{The probability to evaluate the time at}
}
\description{
This numerically evaluates the inverse-CDF function of the Curve object at the specified p
}
\examples{
evaluateInvfunction(Weibull(100,1), 0.5)
}
