% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{evaluateInvfunction}
\alias{evaluateInvfunction}
\title{Method for evaluating the inverse-CDF function for a Curve object at p}
\usage{
evaluateInvfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{...}{Pass-through arguments}
}
\description{
This numerically evaluates the inverse-CDF function of the Curve object at the specified p
}
\examples{
evaluateInvfunction(Weibull(100,1), 0.5)
}
