% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{evaluatePDFfunction,Curve-method}
\alias{evaluatePDFfunction,Curve-method}
\title{Method for evaluating the PDF function for a Curve object at x}
\usage{
\S4method{evaluatePDFfunction}{Curve}(theObject, x)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{x}{The time to evaluate at}
}
\description{
This numerically evaluates the PDF function of the Curve object at the specified x
}
\examples{
evaluatePDFfunction(Weibull(100,1),10)
}
