% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{evaluatePDFfunction}
\alias{evaluatePDFfunction}
\title{Method for evaluating the PDF function for a Curve object at x}
\usage{
evaluatePDFfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{...}{Pass-through arguments}
}
\description{
This numerically evaluates the PDF function of the Curve object at the specified x
}
\examples{
evaluatePDFfunction(Weibull(100,1),10)
}
