% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_functions.R
\name{events2power}
\alias{events2power}
\title{Calculate Schoenfeld or Event Proportion power from number of events}
\usage{
events2power(events, HR, ratio = 1, alpha1 = 0.025)
}
\arguments{
\item{events}{Number of events.}

\item{HR}{Hazard Ratio. Values below 1 indicate a benefit to the active arm vs control.}

\item{ratio}{Randomisation ratio (Schoenfeld), or event ratio (event proportion method). Default=1.}

\item{alpha1}{1-sided alpha. Default=0.025.}
}
\value{
Power as a decimal
}
\description{
Calculate Schoenfeld or Event Proportion power from number of events
}
\keyword{internal}
