% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_constructors.R
\name{Exponential}
\alias{Exponential}
\title{Exponential Curve constructor function}
\usage{
Exponential(lambda)
}
\arguments{
\item{lambda}{Rate parameter for Exponential distribution.}
}
\description{
This creates a Curve object for a Exponential distribution.\cr
Curve objects contain all necessary information to describe a distribution, including functions and parameters describing it.\cr
Parameterisation follows that used by pexp etc. Note that lambda = 1/alpha from the Weibull constructor. See Details for more information on parameterisation.
}
\details{
The exponential distribution has parameterisation:\cr
f(x) = lambda e^(- lambda x)\cr
F(x) = 1 - e^(- lambda x)
}
\examples{
Exponential(0.01)
}
\author{
James Bell
}
