% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_functions.R
\name{freedmanpower}
\alias{freedmanpower}
\title{Calculate Freedman power from number of events}
\usage{
freedmanpower(events, HR, ratio = 1, alpha1 = 0.025)
}
\arguments{
\item{events}{Number of events.}

\item{HR}{Hazard Ratio.}

\item{ratio}{Randomisation ratio. Default=1.}

\item{alpha1}{1-sided alpha. Default=0.025.}
}
\value{
Power as a decimal
}
\description{
Calculate Freedman power from number of events
}
\keyword{internal}
