% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getAssessCDFfunction,RCurve-method}
\alias{getAssessCDFfunction,RCurve-method}
\title{Method for returning the CDF function for a RCurve object}
\usage{
\S4method{getAssessCDFfunction}{RCurve}(theObject, q = "q")
}
\arguments{
\item{theObject}{The name of the RCurve Object}

\item{q}{The parameter name to use in the CDF function. Default=q}
}
\description{
This retrieves the CDF function of the specified RCurve object as a string, writing in the assessment time parameter
}
\examples{
getAssessCDFfunction(LinearR(12,100,100))
}
