% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getAssessCDFfunction}
\alias{getAssessCDFfunction}
\title{Method for returning the CDF function for a RCurve object}
\usage{
getAssessCDFfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the RCurve Object}

\item{...}{Pass-through arguments}
}
\description{
This retrieves the CDF function of the specified RCurve object as a string, writing in the assessment time parameter
}
\examples{
getAssessCDFfunction(LinearR(12,100,100))
}
