% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getCDFfunction,Curve-method}
\alias{getCDFfunction,Curve-method}
\title{Method for returning the CDF function for a Curve object}
\usage{
\S4method{getCDFfunction}{Curve}(theObject, q = "q")
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{q}{The parameter name to use in the CDF function. Default=q}
}
\description{
This retrieves the CDF function of the specified Curve object as a string
}
\examples{
getCDFfunction(Weibull(100,1))
}
