% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getInvfunction,Curve-method}
\alias{getInvfunction,Curve-method}
\title{Method for returning the inverse-CDF function for a Curve object}
\usage{
\S4method{getInvfunction}{Curve}(theObject, p = "p")
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{p}{The probability parameter name to use in the CDF function. Default=p}
}
\description{
This retrieves the inverse CDF function of the specified Curve object as a string
}
\examples{
getCDFfunction(Weibull(100,1))
}
