% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getInvfunction}
\alias{getInvfunction}
\title{Method for returning the inverse-CDF function for a Curve object}
\usage{
getInvfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{...}{Pass-through arguments}
}
\description{
This retrieves the full inverse-CDF function of the Curve object as a string
}
\examples{
getInvfunction(Weibull(100,1))
}
