% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getLength,RCurve-method}
\alias{getLength,RCurve-method}
\title{Method for returning the recruitment length from a RCurve}
\usage{
\S4method{getLength}{RCurve}(theObject)
}
\arguments{
\item{theObject}{The name of the RCurve Object}
}
\description{
This returns the RCurve recruitment length
}
\examples{
getLength(LinearR(12,100,100))
}
