% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getMaxF}
\alias{getMaxF}
\title{Method for returning maximum duration of patient follow-up from a RCurve}
\usage{
getMaxF(theObject)
}
\arguments{
\item{theObject}{The name of the RCurve Object}
}
\description{
This returns the RCurve maximum patient follow-up
}
\examples{
getMaxF(LinearR(12,100,100))
}
