% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{getN,RCurve-method}
\alias{getN,RCurve-method}
\title{Method for returning the total patient number from a RCurve}
\usage{
\S4method{getN}{RCurve}(theObject)
}
\arguments{
\item{theObject}{The name of the RCurve Object}
}
\description{
This returns the RCurve total patient number
}
\examples{
getN(LinearR(12,100,100))
}
