% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getN}
\alias{getN}
\title{Method for returning the total patient number from a RCurve}
\usage{
getN(theObject)
}
\arguments{
\item{theObject}{The name of the RCurve Object}
}
\description{
This returns the RCurve total patient number
}
\examples{
getN(LinearR(12,100,100))
}
