% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getNactive}
\alias{getNactive}
\title{Method for returning the active arm patient number from a RCurve}
\usage{
getNactive(theObject)
}
\arguments{
\item{theObject}{The name of the RCurve Object}
}
\description{
This returns the RCurve active arm patient number
}
\examples{
getNactive(LinearR(12,100,100))
}
