% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{getNames,Curve-method}
\alias{getNames,Curve-method}
\title{Method for returning all parameter names from a Curve object}
\usage{
\S4method{getNames}{Curve}(theObject)
}
\arguments{
\item{theObject}{The name of the Curve Object}
}
\description{
This retrieves all parameter names from a Curve object
}
\examples{
getNames(Weibull(100,1))
}
