% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getNames}
\alias{getNames}
\title{Method for returning all parameter names from a Curve object}
\usage{
getNames(theObject)
}
\arguments{
\item{theObject}{The name of the Curve Object}
}
\description{
This retrieves all parameter names from a Curve object
}
\examples{
getNames(Weibull(100,1))
}
