% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{getNcontrol,RCurve-method}
\alias{getNcontrol,RCurve-method}
\title{Method for returning the control arm patient number from a RCurve}
\usage{
\S4method{getNcontrol}{RCurve}(theObject)
}
\arguments{
\item{theObject}{The name of the RCurve Object}
}
\description{
This returns the RCurve control arm patient number
}
\examples{
getNcontrol(LinearR(12,100,100))
}
