% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getNcontrol}
\alias{getNcontrol}
\title{Method for returning the control arm patient number from a RCurve}
\usage{
getNcontrol(theObject)
}
\arguments{
\item{theObject}{The name of the RCurve Object}
}
\description{
This returns the RCurve control arm patient number
}
\examples{
getNcontrol(LinearR(12,100,100))
}
