% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{getParam,Curve-method}
\alias{getParam,Curve-method}
\title{Method for returning a single parameter from a Curve object}
\usage{
\S4method{getParam}{Curve}(theObject, param)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{param}{The number of the parameter that is required}
}
\description{
This retrieves a single parameter from a Curve object
}
\examples{
getParam(Weibull(100,1),1)
}
