% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getParams}
\alias{getParams}
\title{Method for returning all parameters from a Curve object as a list}
\usage{
getParams(theObject)
}
\arguments{
\item{theObject}{The name of the Curve Object}
}
\description{
This retrieves all parameters from a Curve object as a list
}
\examples{
getParams(Weibull(100,1))
}
