% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{getParamsV,Curve-method}
\alias{getParamsV,Curve-method}
\title{Method for returning all parameters from a Curve object as a vector}
\usage{
\S4method{getParamsV}{Curve}(theObject)
}
\arguments{
\item{theObject}{The name of the Curve Object}
}
\description{
This retrieves all parameters from a Curve object as a vector
}
\examples{
getParamsV(Weibull(100,1))
}
