% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getParamsV}
\alias{getParamsV}
\title{Method for returning all parameters from a Curve object as a vector}
\usage{
getParamsV(theObject)
}
\arguments{
\item{theObject}{The name of the Curve Object}
}
\description{
This retrieves all parameters from a Curve object as a vector
}
\examples{
getParamsV(Weibull(100,1))
}
