% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getPatients,RCurve-method}
\alias{getPatients,RCurve-method}
\title{Method for calculating expected number of recruited patients at a given time from an RCurve}
\usage{
\S4method{getPatients}{RCurve}(theObject, x)
}
\arguments{
\item{theObject}{The name of the RCurve Object}

\item{x}{time at which to calculate expected patients recruited"}
}
\description{
This calculates the expected number of recruited patients at a given time based upon an RCurve
}
\examples{
getPatients(LinearR(12,100,100),7)
}
