% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{getPDFfunction,Curve-method}
\alias{getPDFfunction,Curve-method}
\title{Method for returning the PDF function for a Curve object}
\usage{
\S4method{getPDFfunction}{Curve}(theObject, x = "x")
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{x}{The parameter name to use in the PDF function. Default=x}
}
\description{
This retrieves the PDF function of the Curve object as a string
}
\examples{
getPDFfunction(Weibull(100,1))
}
