% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getPDFfunction}
\alias{getPDFfunction}
\title{Method for returning the PDF function for a Curve object}
\usage{
getPDFfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{...}{Pass-through arguments}
}
\description{
This retrieves the PDF function of the Curve object as a string
}
\examples{
getPDFfunction(Weibull(100,1))
}
