% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getRatio,RCurve-method}
\alias{getRatio,RCurve-method}
\title{Method for returning the recruitment ratio from a RCurve}
\usage{
\S4method{getRatio}{RCurve}(theObject)
}
\arguments{
\item{theObject}{The name of the RCurve Object}
}
\description{
This returns the RCurve recruitment ratio
}
\examples{
getRatio(LinearR(12,100,100))
}
