% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getRFfunction}
\alias{getRFfunction}
\title{Method for returning the RF function for a Curve object}
\usage{
getRFfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{...}{Pass-through arguments}
}
\description{
This retrieves the full RF function of the Curve object as a string
}
\examples{
getRFfunction(Weibull(100,1))
}
