% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{getType,Curve-method}
\alias{getType,Curve-method}
\title{Method for returning the Curve type}
\usage{
\S4method{getType}{Curve}(theObject)
}
\arguments{
\item{theObject}{The name of the Curve Object}
}
\description{
This returns the Curve object type
}
\examples{
getType(Weibull(100,1))
}
