% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_constructors.R
\name{LinearR}
\alias{LinearR}
\title{LinearR RCurve constructor function}
\usage{
LinearR(rlength, Nactive, Ncontrol)
}
\arguments{
\item{rlength}{Length of recruitment.}

\item{Nactive}{Number of patients recruited in active arm.}

\item{Ncontrol}{Number of patients recruited in active arm.}
}
\description{
This creates a RCurve object for a linear recruitment distribution.\cr
RCurve objects contain all necessary information to describe a recruitment distribution. They are a particular type of Curve object containing additional recruitment-related information, including patient numbers and the randomisation ratio.\cr
}
\details{
This RCurve is used when it is expected that patients enter a trial at a constant rate until the required number is achieved.
}
\examples{
LinearR(rlength=12,Nactive=100,Ncontrol=100)
}
\author{
James Bell
}
