% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{plotSF,Curve-method}
\alias{plotSF,Curve-method}
\title{Method for plotting the Survival Function of a Curve object}
\usage{
\S4method{plotSF}{Curve}(theObject, overlay = FALSE, maxT = 100,
  increment = 0.1, xlab = "Time", ylab = "S(t)",
  main = "Kaplan Meier plot", type = "l", ...)
}
\arguments{
\item{theObject}{The name of the RCurve Object}

\item{overlay}{Boolean whether to overlay on existing one (vs start a new one). Default=FALSE}

\item{maxT}{Maximum time to plot up to. Default=100}

\item{increment}{Plotting time increment. Default=0.1}

\item{xlab}{X-axis label. Default="Time"}

\item{ylab}{Y-axis label. Default="S(t)"}

\item{main}{title of plot. Default="Kaplan Meier plot"}

\item{type}{type of plot (see standard graphical parameters). Default="l" (lines).}

\item{...}{Standard graphical parameter arguments to be passed on to 'plot'/'lines', e.g. to change appearance of plot.}
}
\description{
This plots a Curve Survival Function
}
\examples{
plotSF(Weibull(100,1))
plotSF(Weibull(100,1),xlab="Test x label",maxT=60)
plotSF(Weibull(80,0.8),overlay=TRUE,col=2,lty=2)
}
