% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_functions.R
\name{power2events}
\alias{power2events}
\title{Calculate Schoenfeld or Event Proportion based event numbers from power}
\usage{
power2events(power, HR, ratio = 1, alpha1 = 0.025)
}
\arguments{
\item{power}{Desired power (decimal).}

\item{HR}{Hazard Ratio. Values below 1 indicate a benefit to the active arm vs control.}

\item{ratio}{Randomisation ratio (Schoenfeld), or event ratio (event proportion method). Default=1.}

\item{alpha1}{1-sided alpha. Default=0.025.}
}
\value{
Required event number
}
\description{
Calculate Schoenfeld or Event Proportion based event numbers from power
}
\keyword{internal}
