% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{random_draw,Curve-method}
\alias{random_draw,Curve-method}
\title{Method for taking random draws from a Curve object distribution}
\usage{
\S4method{random_draw}{Curve}(theObject, n)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{n}{Number of random draws required}
}
\description{
This takes random draws from the Curve object distribution
Note that the seed should always be appropriately set before invoking this
}
\examples{
random_draw(Weibull(100,1),1000)
}
