% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_gestate}
\alias{run_gestate}
\title{Load Shiny for Gestate
Loads the Shiny interactive GUI for gestate}
\usage{
run_gestate()
}
\description{
Load Shiny for Gestate
Loads the Shiny interactive GUI for gestate
}
\examples{
\donttest{run_gestate()}
}
