% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{setPatients,RCurve-method}
\alias{setPatients,RCurve-method}
\title{Method for setting N's in an RCurve}
\usage{
\S4method{setPatients}{RCurve}(theObject, Nactive, Ncontrol)
}
\arguments{
\item{theObject}{The name of the RCurve Object}

\item{Nactive}{Number of patients to set in active arm}

\item{Ncontrol}{Number of patients to set in control arm}
}
\description{
This sets the RCurve patient numbers per arm directly and updates N and Ratio accordingly
}
\examples{
setPatients(LinearR(12,100,100),200,150)
}
