% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{setPatients}
\alias{setPatients}
\title{Method for setting N's in an RCurve}
\usage{
setPatients(theObject, ...)
}
\arguments{
\item{theObject}{The name of the RCurve Object}

\item{...}{Pass-through arguments}
}
\description{
This sets the RCurve patient numbers per arm directly and updates N and Ratio accordingly
}
\examples{
setPatients(LinearR(12,100,100),200,150)
}
