% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{show,Curve-method}
\alias{show,Curve-method}
\title{Method for displaying Curve objects neatly - replaces standard show method}
\usage{
\S4method{show}{Curve}(object)
}
\arguments{
\item{object}{The name of the Curve Object}
}
\description{
This is the standard print method for a Curve object
}
\examples{
Weibull(100,1)
}
