% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{show,RCurve-method}
\alias{show,RCurve-method}
\title{Method for displaying RCurve objects neatly - replaces standard show method}
\usage{
\S4method{show}{RCurve}(object)
}
\arguments{
\item{object}{The name of the RCurve Object}
}
\description{
This is the standard print method for an RCurve Object
}
\examples{
LinearR(12,100,100)
}
