% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_trials_strata}
\alias{simulate_trials_strata}
\title{Perform multi-strata simulations of time-to-event data using arbitrary event, censoring and recruitment distributions.}
\usage{
simulate_trials_strata(
  stratum_probs,
  active_ecurve,
  control_ecurve,
  active_dcurve = Blank(),
  control_dcurve = Blank(),
  rcurve,
  assess = NULL,
  fix_events = NULL,
  stratum_name = "Stratum",
  iterations,
  seed,
  detailed_output = FALSE,
  output_type = c("matrix", "list")
)
}
\arguments{
\item{stratum_probs}{Vector of probabilities that patients belong to each stratum. Must sum to 1. Its length determines the number of strata.}

\item{active_ecurve}{List of event distributions for the active arm, specified as a list of Curve objects. If single Curve is specified, will be used for all strata.}

\item{control_ecurve}{List of event distributions for the control arm, specified as a list of Curve objects. If single Curve is specified, will be used for all strata.}

\item{active_dcurve}{List of dropout/censoring distribution for the active arm, specified as a Curve object. If single Curve is specified, will be used for all strata. By default, Blank(), i.e. no dropout in any stratum.}

\item{control_dcurve}{List of dropout/censoring distribution for the control arm, specified as a Curve object. If single Curve is specified, will be used for all strata. By default, Blank(), i.e. no dropout in any stratum.}

\item{rcurve}{Recruitment distribution, specified as a single RCurve object.}

\item{assess}{Positive number for the assessment time at which administrative censoring will be performed.}

\item{fix_events}{Positive integer for the number of events to fix (if required), letting the assessment time vary. Alternatively, NULL for fixed time assessment with variable event numbers. Notes: Fixing event numbers overrides any specified assessment time and slows simulation considerably. Default = NULL (fixed analysis time)}

\item{stratum_name}{Name of the column defining the stratum. Default="Stratum".}

\item{iterations}{Number of simulations to perform. Depending on trial size, 10,000-20,000 is typically OK to analyse on 8GB RAM.}

\item{seed}{Seed number to use. Numerical, although if "Rand" is specified, a system-time-derived number will be used.}

\item{detailed_output}{Boolean to require full details of timings of competing processes. If FALSE, the simplified data only includes the *'ed output columns - this approximately halves RAM requirements. Default=FALSE (simplified).}

\item{output_type}{"matrix" or "list" specifying the type of output required. "matrix" requests a single matrix with a column "iter" to denote the simulation, while "list" creates a list with one entry per simulation. Default="matrix".}
}
\value{
Returns a table with one row per patient per simulation. Table contains the following columns:
\itemize{
 \item{"Time"}{ Simulated actually observed (patient) time of event or censoring: This is the main column of interest for analysis*}
 \item{"Censored"}{ Simulated censoring indicator: 1 denotes censoring (administrative or dropout), 0 denotes an event*}
 \item{"Trt"}{ Treatment group number - 1 is active, 2 is control*}
 \item{"Iter"}{ Simulation number*}
 \item{"ETime"}{ Simulated actual event (patient) time (may or may not be observed)}
 \item{"CTime"}{ Simulated actual censoring/dropout (patient) time (may or may not be observed)}
 \item{"Rec_Time"}{ Simulated (trial) time of recruitment}
 \item{"Assess"}{ Prespecified (trial) time of assessment}
 \item{"RCTime"}{ Simulated actual administrative censoring (patient) time (may or may not be observed)}
 \item{"Stratum"}{ Stratum number. Column name will be the value of the stratum_name argument.)}
}
}
\description{
Function for simulating generalised two-arm multi-strata time-to-event trial data for NPH trials with arbitrary event, censoring and recruitment distributions.\cr
Acts as a wrapper for simulate_trials.\cr
Vector of strata proportions supplies number of strata. Event and censoring distributions specified via lists of Curve objects. If only one Curve supplied then assumed to be common to all strata. Recruitment specified via a single RCurve object.\cr
As it uses same architecture and similar syntax to nph_traj(), results in simple cases may be directly comparable to e.g. use of MixExp() or MixWei() Curves.\cr
Can be used to validate outputs from nph_traj().\cr
Data sets from this are set up to be automatically analysed with the analyse_sim function (including stratified analysis if you provide it the name of stratum column).\cr
}
\examples{
example_strat_sim <- simulate_trials_strata(stratum_probs=c(0.5,0.5),
active_ecurve=c(Weibull(250,0.8),Weibull(100,1)), control_ecurve=Weibull(100,1),
rcurve=LinearR(12,100,100),assess=20,iterations=5,seed=12345)
}
\author{
James Bell
}
