% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_functions.R
\name{ZV2power}
\alias{ZV2power}
\title{Calculate Normal distribution test power based on Effect size and Variance}
\usage{
ZV2power(Z, V = 1, alpha1 = 0.025)
}
\arguments{
\item{Z}{effect size.}

\item{V}{variance. Default=1.}

\item{alpha1}{1-sided alpha. Default=0.025.}
}
\value{
Power as a decimal
}
\description{
Calculate Normal distribution test power based on Effect size and Variance
}
\keyword{internal}
