% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\docType{methods}
\name{createRFfunction,Curve-method}
\alias{createRFfunction,Curve-method}
\title{Method for running the RF function for a Curve object}
\usage{
\S4method{createRFfunction}{Curve}(theObject, n = "n")
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{n}{The parameter name for the number of random draws to make. Default=n}
}
\description{
This retrieves and runs the full RF function of the Curve object as a string
}
\examples{
createRFfunction(Weibull(100,1))
}
