% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_architecture.R
\name{createRFfunction}
\alias{createRFfunction}
\title{Method for creating a random draw function from a Curve object}
\usage{
createRFfunction(theObject, ...)
}
\arguments{
\item{theObject}{The name of the Curve Object}

\item{...}{Pass-through arguments}
}
\description{
This creates a random draw function from the Curve object
}
\examples{
createRFfunction(Weibull(100,1))
}
